+++ src/headless/converse-core.js
@@ -104,7 +104,7 @@ const _converse = {
     'promises': {}
 }
 
-_converse.VERSION_NAME = "v5.0.1dev";
+_converse.VERSION_NAME = "v5.0.2dev";
 
 Object.assign(_converse, Backbone.Events);
 
@@ -115,6 +115,14 @@ _converse.Collection = Backbone.Collection.extend({
         this.reset();
     }
 });
+
+
+/**
+ * Custom error for indicating timeouts
+ * @namespace _converse
+ */
+class TimeoutError extends Error {}
+_converse.TimeoutError = TimeoutError;
 
 
 // Make converse pluggable
@@ -1455,16 +1463,18 @@ _converse.api = {
      * {@link _converse.api.listen.on} or {@link _converse.api.listen.once}
      * (see [_converse.api.listen](http://localhost:8000/docs/html/api/-_converse.api.listen.html)).
      *
+     * Some events also double as promises and can be waited on via {@link _converse.api.waitUntil}.
+     *
      * @method _converse.api.trigger
      * @param {string} name - The event name
      * @param {...any} [argument] - Argument to be passed to the event handler
      * @param {object} [options]
      * @param {boolean} [options.synchronous] - Whether the event is synchronous or not.
-     *    When a synchronous event is fired, Converse will wait for all
-     *    promises returned by the event's handlers to finish before continuing.
+     *  When a synchronous event is fired, a promise will be returned
+     *  by {@link _converse.api.trigger} which resolves once all the
+     *  event handlers' promises have been resolved.
      */
     async trigger (name) {
-        /* Event emitter and promise resolver */
         const args = Array.from(arguments);
         const options = args.pop();
         if (options && options.synchronous) {
@@ -1921,8 +1931,8 @@ const converse = {
      * @memberOf converse
      */
     plugins: {
-        /** Registers a new plugin.
-         *
+        /**
+         * Registers a new plugin.
          * @method converse.plugins.add
          * @param {string} name The name of the plugin
          * @param {object} plugin The plugin object