+++ src/utils/html.js
@@ -161,33 +161,29 @@ u.renderImageURL = function (_converse, url) {
 };
 
 
+/**
+ * Returns a Promise which resolves once all images have been loaded.
+ * @returns {Promise}
+ */
 u.renderImageURLs = function (_converse, el) {
-    /* Returns a Promise which resolves once all images have been loaded.
-     */
     if (!_converse.show_images_inline) {
         return Promise.resolve();
     }
-    const { __ } = _converse;
     const list = el.textContent.match(URL_REGEX) || [];
     return Promise.all(
-        _.map(list, url =>
+        list.map(url =>
             new Promise((resolve, reject) => {
                 if (u.isImageURL(url)) {
                     return isImage(url).then(img => {
                         const i = new Image();
                         i.src = img.src;
                         i.addEventListener('load', resolve);
-                        // We also resolve for non-images, otherwise the
-                        // Promise.all resolves prematurely.
+                        // We also resolve (instead of reject) for non-images,
+                        // otherwise the Promise.all resolves prematurely.
                         i.addEventListener('error', resolve);
-
                         const { __ } = _converse;
-                        _.each(sizzle(`a[href="${url}"]`, el), (a) => {
-                            a.outerHTML= tpl_image({
-                                'url': url,
-                                'label_download': __('Download')
-                            })
-                        });
+                        sizzle(`a[href="${url}"]`, el)
+                            .forEach(a => (a.outerHTML = tpl_image({url, 'label_download': __('Download')})));
                     }).catch(resolve)
                 } else {
                     return resolve();
@@ -227,7 +223,7 @@ u.calculateElementHeight = function (el) {
 
 u.getNextElement = function (el, selector='*') {
     let next_el = el.nextElementSibling;
-    while ((next_el instanceof Element) && !sizzle.matchesSelector(next_el, selector)) {
+    while (next_el !== null && !sizzle.matchesSelector(next_el, selector)) {
         next_el = next_el.nextElementSibling;
     }
     return next_el;
@@ -235,24 +231,24 @@ u.getNextElement = function (el, selector='*') {
 
 u.getPreviousElement = function (el, selector='*') {
     let prev_el = el.previousElementSibling;
-    while ((prev_el instanceof Element) && !sizzle.matchesSelector(prev_el, selector)) {
-        prev_el = prev_el.previousSibling
+    while (prev_el !== null && !sizzle.matchesSelector(prev_el, selector)) {
+        prev_el = prev_el.previousElementSibling
     }
     return prev_el;
 }
 
 u.getFirstChildElement = function (el, selector='*') {
     let first_el = el.firstElementChild;
-    while ((first_el instanceof Element) && !sizzle.matchesSelector(first_el, selector)) {
-        first_el = first_el.nextSibling
+    while (first_el !== null && !sizzle.matchesSelector(first_el, selector)) {
+        first_el = first_el.nextElementSibling
     }
     return first_el;
 }
 
 u.getLastChildElement = function (el, selector='*') {
     let last_el = el.lastElementChild;
-    while ((last_el instanceof Element) && !sizzle.matchesSelector(last_el, selector)) {
-        last_el = last_el.previousSibling
+    while (last_el !== null && !sizzle.matchesSelector(last_el, selector)) {
+        last_el = last_el.previousElementSibling
     }
     return last_el;
 }
@@ -288,7 +284,7 @@ u.hideElement = function (el) {
 
 u.ancestor = function (el, selector) {
     let parent = el;
-    while ((parent instanceof Element) && !sizzle.matchesSelector(parent, selector)) {
+    while (parent !== null && !sizzle.matchesSelector(parent, selector)) {
         parent = parent.parentElement;
     }
     return parent;
@@ -298,7 +294,7 @@ u.nextUntil = function (el, selector, include_self=false) {
     /* Return the element's siblings until one matches the selector. */
     const matches = [];
     let sibling_el = el.nextElementSibling;
-    while ((sibling_el instanceof Element) && !sibling_el.matches(selector)) {
+    while (sibling_el !== null && !sibling_el.matches(selector)) {
         matches.push(sibling_el);
         sibling_el = sibling_el.nextElementSibling;
     }